+++ lib/wd_sinatra.rb
@@ -1,7 +1,9 @@
 require "wd_sinatra/version"
+require 'wd_sinatra/app_loader'
+require 'wd_sinatra/sinatra_ext'
+require 'wd_sinatra/test_helpers'
+require 'wd_sinatra/test_unit_helpers'
+require 'wd_sinatra/ws_list_ext'
 
-module WD
-  module Sinatra
-    # Your code goes here...
-  end
+module WDSinatra
 end